backend/ml_models/requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
tensorflow==2.15.0
pymongo==4.6.1
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
joblib==1.3.2
